backend/requirements.txt
# Phase 8: Real-Time Intelligence with Human-in-the-Loop
# Core dependencies for the hybrid intelligence system

# Twilio for SMS and phone call alerts
twilio>=8.0.0

# Desktop notifications
plyer>=2.1.0

# YAML configuration parsing
PyYAML>=6.0

# CLI table formatting
tabulate>=0.9.0

# Numerical operations for feedback analysis
numpy>=1.21.0

# Async HTTP requests (if needed for webhooks)
aiohttp>=3.8.0

# Date/time utilities
python-dateutil>=2.8.0

# JSON schema validation (optional)
jsonschema>=4.0.0

# Logging utilities
coloredlogs>=15.0

# Testing dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development dependencies (optional)
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0